feat(department): update department management

+++ backend/app/src/main/java/com/hustleflow/department/service/DepartmentService.java
@@ -2,6 +2,7 @@ package com.hustleflow.department.service;
 
 import java.util.List;
 import java.util.Optional;
+import java.util.stream.Collectors;
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -9,9 +10,12 @@ import org.springframework.stereotype.Service;
 
 import com.hustleflow.department.domain.Department;
 import com.hustleflow.department.dto.request.DepartmentRequestDTO;
+import com.hustleflow.department.dto.response.ManagerSummaryDto;
+import com.hustleflow.department.dto.response.DeparmentResponseDTO;
 import com.hustleflow.department.repository.DepartmentRepository;
 import com.hustleflow.employee.domain.Employee;
 import com.hustleflow.employee.repository.EmployeeRepository;
+import com.hustleflow.exception.ResourceNotFoundException;
 
 @Service
 public class DepartmentService {
@@ -26,13 +30,14 @@ public class DepartmentService {
         this.employeeRepository = employeeRepository;
     }
 
-    public List<Department> getDepartments() {
-        return departmentRepository.findAll();
+    public List<DeparmentResponseDTO> getDepartments() {
+        List<Department> departments = departmentRepository.findAll();
+
+        return departments.stream().map(this::mapToResponse).collect(Collectors.toList());
     }
 
     public String createDepartment(DepartmentRequestDTO request) {
-        Employee manager = employeeRepository.findById(request.getManagerId())
-                .orElseThrow(() -> new RuntimeException("Manager doesn't exist: " + request.getManagerId()));
+        Employee manager = employeeRepository.findById(request.getManagerId()).orElse(null);
         Optional<Department> existing = departmentRepository.findByDepartmentName(request.getDepartmentName());
         existing.ifPresent(obj -> {
             throw new RuntimeException("Department already exists: " + obj.getDepartmentName());
@@ -72,12 +77,32 @@ public class DepartmentService {
     public String deleteDepartment(Long id) {
         Optional<Department> existing = departmentRepository.findById(id);
         if (existing.isEmpty()) {
-            throw new RuntimeException("Department doesn't exists: " + id);
+            throw new ResourceNotFoundException("Department doesn't exists: " + id);
+        }
+
+        if (existing.get().getManager() != null) {
+            throw new RuntimeException("Cannot remove department as it is holding a reference to another object");
         }
 
         departmentRepository.deleteById(id);
         log.info("Department removed: " + id);
         return "Department has been deleted";
+    }
+
+    private DeparmentResponseDTO mapToResponse(Department department) {
+        DeparmentResponseDTO dto = DeparmentResponseDTO.builder()
+                .id(department.getId())
+                .departmentName(department.getDepartmentName())
+                .code(department.getCode())
+                .description(department.getDescription()).build();
+
+        if (department.getManager() != null) {
+            ManagerSummaryDto managerSummary = ManagerSummaryDto.builder().id(department.getManager().getId())
+                    .name(department.getManager().getName()).age(department.getManager().getAge()).build();
+            dto.setManager(managerSummary);
+        }
+
+        return dto;
     }
 
 }
